requirements.txt
argon2-cffi >= 23.1.0
Flask >= 3.0.2
flask-sqlalchemy >= 3.1.1
psycopg2-binary >= 2.9.9
python-dotenv >= 1.0.1

# Linters etc
pytest >= 8.1.1
coverage >= 7.4.4
pytest-cov >= 4.1.0
mypy >= 1.9.0
pylint >= 3.1.0

# Development
lorem >= 0.1.1